+++ index.html
@@ -453,5 +453,5 @@
       </div>
     </template>
   </body>
-  <script src="js/functions.js"></script>
+  <script src="js/main.js" type="module"></script>
 </html>
