+++ js/main.js
@@ -1,0 +1,3 @@
+import "./functions";
+import "./util";
+import "./data";
