+++ js/util.js
@@ -1,0 +1,6 @@
+// нужна функция которая будет возвращать случайное число в заданном диапазоне
+function getRandomIntInclusive(min, max) {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1) + min); // Максимум и минимум включаются
+}
